% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decompose_method.R
\name{decompose}
\alias{decompose}
\title{A Decompose Time Series Method}
\usage{
decompose(data, supply, key, year, qtr_key)
}
\arguments{
\item{data:}{Complete data in a data frame or a data table or a tibbe which needs to be decomposed}

\item{supply:}{Target attribute on which decomposition should happen}

\item{key:}{Key attributes in the table on which split should happen. For example: state_id, job_code_id}

\item{year:}{Year attribute of the data frame.}

\item{qtr_key:}{Attribute that contains the quarter value.}
}
\value{
A data frame with attribute:\cr
trend: Trend component of the data\cr
seasonal: Seasonal component of the data\cr
remainder: Random Noise of the time series
}
\description{
User calls this method to decompose the time series into trend, seasonality and random noise.
The function will check if all the quarters in each key has atleast a value; if not it will throw an error.
The function will split the complete data according to its key and call decompose_ts() to decompose the time series.
}
