% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decompose_method.R
\name{decompose_ts}
\alias{decompose_ts}
\title{Decompose the time series}
\usage{
decompose_ts(data, supply, year)
}
\arguments{
\item{data:}{A data frame or data table with data of just one series}

\item{supply:}{Target attribute on which decomposition should happen}

\item{year:}{Year attribute of the data frame.}
}
\description{
This function is called once made sure that all quarters have atleast one data point for a series. 
It will be called by another method decompose() once all the checks are done.
It will first convert the target column into time series and then split it into:
Trend, 
Seasonality and
Random Noise\cr
NOTE: The function is not exported, because it should not be called.
}
