% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grubbs_test_cpp}
\alias{grubbs_test_cpp}
\title{Grubbs' Test}
\usage{
grubbs_test_cpp(y, n, alpha)
}
\arguments{
\item{y}{NumericVector}

\item{n}{IntegerVector}

\item{alpha}{double}
}
\description{
Grubbs' Test
}

