% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{imputation}
\alias{imputation}
\title{Missing value imputation for the dataset with sparse data}
\usage{
imputation(complete, incomp, digit_coding = NA, digit_colnames = NA,
  occ_code, key, compl_supply, incomp_supply, qtr, year)
}
\arguments{
\item{complete:}{Th complete data frame with the imputated values from ts impute}

\item{incomp:}{The data that needs to be imputed.}

\item{digit_coding:}{A dataset that will contain all th elevel of job occupation. 
If NA, then the function will assume that 1level code is substring of 1 digit in occupation code, 
2 level code is substring of first 2 digits and so on. DEFAULT to NA}

\item{digit_colnames:}{The column name of the 1 level code, 2 level code and so on in digit_coding dataset. 
Please maintain the chronological order}

\item{occ_code:}{Attribute that contains occupational code}

\item{key:}{Attributes that act as the key. Note: No need to add occupational code in this.}

\item{compl_supply:}{The column name of the supply column in complete dataset}

\item{incomp_supply:}{The column name of the supply column in incomplete dataset that needs to be imputed}

\item{qtr:}{Attribute with quarter value}

\item{year:}{Attribute with year value}
}
\value{
The dataset with both the datasets and the supply columns imputed.
}
\description{
In this function, missing data will be imputated for the sparse dataset using the growth rate.
The dataset on which outlier detection is done should be imputed using ts impute. 
That will serve as the complete dataset for the imputation of the next dataset.
}
