% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rolling_method.R
\name{mad_outliers}
\alias{mad_outliers}
\title{MAD Outlier Function}
\usage{
mad_outliers(data, cutoff, replace, mad_const)
}
\arguments{
\item{data:}{Target attribute on which MAD should be calculated}

\item{cutoff:}{The cutoff values against which the deviation will be checked. 
If the computed ratio is less than the cutoff, then the point if not an outlier.}

\item{mad_const:}{Scale factor to calculate MAD values to maintain consistency.}

\item{replace:}{In the final column, the outliers will be replaced by this variable's value.}
}
\description{
This function just calculates the mad ratios and return the replaced series.
 It is called by rollingout() to get the results.\cr
NOTE: The function is not exported, because it should not be called.
}
