% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overall_Method.R
\name{overall_outlier}
\alias{overall_outlier}
\title{Overall MAD Outlier Detection Technique}
\usage{
overall_outlier(data, supply, key, year, qtr_key, cutoff = 1.5,
  replace = NA, mad_const = 1.4826)
}
\arguments{
\item{data:}{Th complete data frame or data table or tibble that contains the data}

\item{supply:}{Target attribute on which decomposition should happen}

\item{key:}{Key attributes in the table on which split should happen. For example: state_id, job_code_id}

\item{year:}{Year attribute of the data frame.}

\item{qtr_key:}{Attribute that contains the quarter value.}

\item{cutoff:}{The cutoff values against which the deviation will be checked. 
If the computed ratio is less than the cutoff, then the point if not an outlier. DEFAULT to 1.5}

\item{mad_const:}{Scale factor to calculate MAD values to maintain consistency. DEFAULT to 1.4826}

\item{replace:}{In the final column, the outliers will be replaced by this variable's value. DEFAULT to NA.}
}
\value{
A data frame with two additional attributes:\cr
spl_without_otl_overallMAD: The target attribute in which all the outliers will be replaced by the replace variable\cr
otl_flag_overallMAD: Outlier flag by overall MAD calculation method. Outliers will be marked as 1.
}
\description{
This funciton is called to get the outlier by overall MAD method.
MAD method will be applied on the complete series of target variable. 
The MAD values will be checked against the cutoff and marked as outlier.\cr
Please note: In this method, the trend and seasonality component is not considered.\cr
This method should never be used alone. One should always ensemble it in some or the other way with other methods.\cr\cr
ADVISE: Do not keep the cutoff too low; lower the cutoff, greater will be the loss in trend.
}
