% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rolling_method.R
\name{rolling_outlier}
\alias{rolling_outlier}
\title{Rooling Outlier Function}
\usage{
rolling_outlier(data, supply, key, year, qtr_key, cutoff = 1.5,
  replace = NA, mad_const = 1.4826, nona = F)
}
\arguments{
\item{data:}{Th complete data frame or data table or tibble that contains the data}

\item{supply:}{Target attribute on which decomposition should happen}

\item{key:}{Key attributes in the table on which split should happen. For example: state_id, job_code_id}

\item{year:}{Year attribute of the data frame.}

\item{qtr_key:}{Attribute that contains the quarter value.}

\item{cutoff:}{The cutoff values against which the deviation will be checked. 
If the computed ratio is less than the cutoff, then the point if not an outlier. DEFAULT to 1.5}

\item{mad_const:}{Scale factor to calculate MAD values to maintain consistency. DEFAULT to 1.4826}

\item{replace:}{In the final column, the outliers will be replaced by this variable's value. DEFAULT to NA.}

\item{nona:}{The variable that can trigger whethere to include NAs in roling technique or not. DEFAULT to False.}
}
\value{
A data frame with two additional attributes:\cr
spl_without_otl_rMAD<_nona>: The target attribute in which all the outliers will be replaced by the replace variable\cr
otl_flag_rMAD<_nona>: Outlier flag by rolling method. Outliers will be marked as 1.
}
\description{
This function is called by the user to get the outlier by rolling methodology including or excluding NAs.
The data set is split on the key and sent to the rollingout() for further processing.
Please note: In this method, if there are less than two data points in a perticular set of 4 quarters, 
then even if the points are outliers they are not marked. 
Precautions should be taken for such cases.\cr
If there are many such cases then this should be emsembled with other methods in some way.
}
