% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rolling_method.R
\name{rollingout}
\alias{rollingout}
\title{Rollinf Outlier Detection technique}
\usage{
rollingout(z, supply, cutoff = 1.5, replace = NA, mad_const = 1.4826,
  nona = F)
}
\arguments{
\item{z:}{The data frame or data table or tibble that contains the data for one time series.}

\item{supply:}{Target attribute on which decomposition should happen.}

\item{cutoff:}{The cutoff values against which the deviation will be checked. 
If the computed ratio is less than the cutoff, then the point if not an outlier. DEFAULT to 1.5}

\item{mad_const:}{Scale factor to calculate MAD values to maintain consistency. DEFAULT to 1.4826}

\item{nona:}{The variable that can trigger whethere to include NAs in roling technique or not. DEFAULT to False.}

\item{replace:}{In the final column, the outliers will be replaced by this variable's value. DEFAULT to NA.}
}
\value{
A data frame with two additional attributes:\cr
spl_without_otl_rMAD<_nona>: The target attribute in which all the outliers will be replaced by the replace variable
otl_flag_rMAD<_nona>: Outlier flag by rolling method. Outliers will be marked as 1.
}
\description{
In this function, a time series for a perticular key is sent. 
This series is then grouped in 4 quarters and rolled by two quarters.
For each group MAD values are calculated.
Apart from the first 2 quarters and last 2 quarters, 
for all the other quarter a point will be qualified as outlier only when if is marked in both the groups.\cr
The same function can be used to find the outliers by rolling method without NAs by triggering the parameter nona.\cr
NOTE: This function is exported, but great care should be taken while using it. 
Only one time series should be sent to the function to get proper results.
}
