% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sequential_ensemble_technique.R
\name{seq_outlier_detection}
\alias{seq_outlier_detection}
\title{Sequential Ensemble Technique to detect Outlier}
\usage{
seq_outlier_detection(data, key, supply, year, qtr_key, seq = c(1, 2, 3,
  4, 5), dec_cutoff = 1.5, dec_mad_const = 1.4826, dec_replace = NA,
  rMAD_cutoff = 1.5, rMAD_mad_const = 1.4826, rMAD_replace = NA,
  rMAD_nona_cutoff = 1.5, rMAD_nona_mad_const = 1.4826,
  rMAD_nona_replace = NA, overall_MAD_cutoff = 1.5,
  overall_MAD_mad_const = 1.4826, overall_MAD_replace = NA,
  runmed_replace = NA, quantile = 0.75)
}
\arguments{
\item{data:}{Th complete data frame or data table or tibble that contains the data}

\item{supply:}{Target attribute on which decomposition should happen}

\item{key:}{Key attributes in the table on which split should happen. For example: state_id, job_code_id}

\item{year:}{Year attribute of the data frame.}

\item{qtr_key:}{Attribute that contains the quarter value.}

\item{seq:}{This variable will get the sequence of the methods to be used. DEFAULT to (1,2,3,4,5)}

\item{dec_cutoff:}{The cutoff values against which the deviation will be checked for decomposition method.
If the computed ratio is less than the cutoff, then the point if not an outlier. DEFAULT to 1.5}

\item{dec_mad_const:}{Scale factor to calculate MAD values in decomposition method to maintain consistency.
DEFAULT to 1.4826}

\item{dec_replace:}{In the final column for decomposition method,
the outliers will be replaced by this variable's value. DEFAULT to NA.}

\item{rMAD_cutoff:}{The cutoff values against which the deviation will be checked for rolling MAD method.
If the computed ratio is less than the cutoff, then the point if not an outlier. DEFAULT to 1.5}

\item{rMAD_mad_const:}{Scale factor to calculate MAD values in rolling MAD method to maintain consistency.
DEFAULT to 1.4826}

\item{rMAD_replace:}{In the final column for rolling MAD method,
the outliers will be replaced by this variable's value. DEFAULT to NA.}

\item{rMAD_nona_cutoff:}{The cutoff values against which the deviation will be checked for rolling MAD method without NAs.
If the computed ratio is less than the cutoff, then the point if not an outlier. DEFAULT to 1.5}

\item{rMAD_nona_mad_const:}{Scale factor to calculate MAD values in rolling MAD method without NAs to maintain consistency.
DEFAULT to 1.4826}

\item{rMAD_nona_replace:}{In the final column for rolling MAD method without NAs,
the outliers will be replaced by this variable's value. DEFAULT to NA.}

\item{overallMAD_cutoff:}{The cutoff values against which the deviation will be checked for overall MAD method.
If the computed ratio is less than the cutoff, then the point if not an outlier. DEFAULT to 1.5}

\item{overallMAD_mad_const:}{Scale factor to calculate MAD values in overall MAD method to maintain consistency.
DEFAULT to 1.4826}

\item{overallMAD_replace:}{In the final column for overall MAD method,
the outliers will be replaced by this variable's value. DEFAULT to NA.}

\item{runmed_replace:}{In the final column for decomposition method,
the outliers will be replaced by this variable's value. DEFAULT to NA.}

\item{runmed_quantile:}{Compute any of 25\%  50\%  75\%  100\% quartile depending on the data requirements}
}
\value{
The dataset with few extra attributes.
If the technique number is not present in the parameter seq, then attributes for that method will not be there.\cr
Technique related attributes will be:\cr
spl_without_otl_dec: The target attribute in which all the outliers will be replaced by the replace variable\cr
otl_flag_dec: Outlier flag by decompose method. Outliers will be marked as 1.\cr
spl_without_otl_rMAD: The target attribute in which all the outliers will be replaced by the replace variable\cr
otl_flag_rMAD: Outlier flag by rolling method. Outliers will be marked as 1.\cr
spl_without_otl_rMAD_nona: The target attribute in which all the outliers will be replaced by the replace variable\cr
otl_flag_rMAD_nona: Outlier flag by rolling method without NAs. Outliers will be marked as 1.\cr
spl_without_otl_overallMAD: The target attribute in which all the outliers will be replaced by the replace variable\cr
otl_flag_overallMAD: Outlier flag by overall MAD calculation method. Outliers will be marked as 1.\cr\cr
spl_without_otl_runmed: The target attribute in which all the outliers will be replaced by the replace variable\cr
otl_flag_runmed: Outlier flag by running median method. Outliers will be marked as 1.\cr
seq_spl_without_otl: The final target variable without the outliers\cr
seq_flag: The final flag to the outliers.\cr
Final variable which will be present in all the files\cr
}
\description{
In this perticular function, all the techniques will be applied in a sequential manner, that means the first method
will be applied on the initial set of target data, then the second method on the ouput of the first tecnique and so on.
The sequence can be given in any combination and permutation.\ The code for the methods are as follows:\cr
1. Decomposiiton Method\cr
2. Rolling Quarter Mehthod\cr
3. Rolling Quarter Mehthod without NAs\cr
4. Overall MAD method.\cr
5. Running Median method \cr\cr
To call any one particular method we can call this
function with with only that methods number in sequence
}
