% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sequential_grubbs_test_cpp}
\alias{sequential_grubbs_test_cpp}
\title{Sequential Grubbs' Tests}
\usage{
sequential_grubbs_test_cpp(in_vector, max_outlier_pct, alpha)
}
\arguments{
\item{in_vector}{NumericVector}

\item{max_outlier_pct}{double}

\item{alpha}{double}
}
\description{
Sequential Grubbs' Tests
}

