% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sequential_grubbs_test_cpp_df}
\alias{sequential_grubbs_test_cpp_df}
\title{Sequential Grubbs' Tests Df}
\usage{
sequential_grubbs_test_cpp_df(df, var_in, var_out, max_outlier_pct, alpha)
}
\arguments{
\item{df}{DataFrame}

\item{var_in}{string}

\item{var_out}{string}

\item{max_outlier_pct}{double}

\item{alpha}{double}
}
\description{
Sequential Grubbs' Tests Df
}

