% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stl_sequential_grubbs_test.R
\name{stl_sequential_grubbs_test}
\alias{stl_sequential_grubbs_test}
\title{STL Sequential Grubbs' Tests}
\usage{
stl_sequential_grubbs_test(series, frequency_season, max_outlier_pct = 0.05,
  alpha = 0.05)
}
\arguments{
\item{series}{numeric vector for time series, can't have NA values}

\item{frequency_season}{string of "minute_day", "hour_day", "hour_week",
"day_week", "day_year", "week_year", "month_year"}

\item{max_outlier_pct}{double}

\item{alpha}{double}
}
\description{
STL Sequential Grubbs' Tests
}

