% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stl_sequential_grubbs_test.R
\name{stl_sequential_grubbs_test_df}
\alias{stl_sequential_grubbs_test_df}
\title{STL Sequential Grubbs' Tests DF}
\usage{
stl_sequential_grubbs_test_df(df, var_in, var_out, frequency_season,
  max_outlier_pct = 0.05, alpha = 0.05)
}
\arguments{
\item{df}{DataFrame}

\item{var_in}{string name of input column, can't have NA values}

\item{var_out}{string name of output column}

\item{frequency_season}{String of "minute_day", "hour_day", "hour_week",
"day_week", "day_year", "week_year", "month_year"}

\item{max_outlier_pct}{double}

\item{alpha}{double}
}
\description{
STL Sequential Grubbs' Tests DF
}

